moved code from pathbuilder to interface.database
