+++ flood_adapt/object_model/hazard/event/eventset.py
@@ -1,7 +1,9 @@
 from pathlib import Path
 from typing import Any
 
-from flood_adapt.dbs_classes.path_builder import ObjectDir
+from flood_adapt.object_model.interface.database import (
+    ObjectDir,
+)
 from flood_adapt.object_model.interface.events import EventSetModel
 from flood_adapt.object_model.interface.object_model import IObject
