+++ flood_adapt/object_model/interface/benefits.py
@@ -5,7 +5,9 @@ from typing import Optional
 import pandas as pd
 from pydantic import BaseModel
 
-from flood_adapt.dbs_classes.path_builder import ObjectDir
+from flood_adapt.object_model.interface.database import (
+    ObjectDir,
+)
 from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
