+++ flood_adapt/object_model/interface/measures.py
@@ -3,7 +3,9 @@ from typing import Any, Generic, Optional, TypeVar
 
 from pydantic import Field, model_validator, validator
 
-from flood_adapt.dbs_classes.path_builder import ObjectDir
+from flood_adapt.object_model.interface.database import (
+    ObjectDir,
+)
 from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
 from flood_adapt.object_model.io.unitfulvalue import (
     UnitfulDischarge,
