+++ flood_adapt/object_model/interface/scenarios.py
@@ -1,6 +1,8 @@
 from abc import abstractmethod
 
-from flood_adapt.dbs_classes.path_builder import ObjectDir
+from flood_adapt.object_model.interface.database import (
+    ObjectDir,
+)
 from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
