+++ flood_adapt/object_model/interface/strategies.py
@@ -1,4 +1,4 @@
-from flood_adapt.dbs_classes.path_builder import ObjectDir
+from flood_adapt.object_model.interface.database import ObjectDir
 from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
