+++ flood_adapt/object_model/strategy.py
@@ -1,10 +1,10 @@
 from typing import Any, Union
 
-from flood_adapt.dbs_classes.path_builder import ObjectDir, db_path
 from flood_adapt.object_model.direct_impact.impact_strategy import ImpactStrategy
 from flood_adapt.object_model.direct_impact.measure.impact_measure import ImpactMeasure
 from flood_adapt.object_model.hazard.hazard_strategy import HazardStrategy
 from flood_adapt.object_model.hazard.measure.hazard_measure import HazardMeasure
+from flood_adapt.object_model.interface.database import ObjectDir, db_path
 from flood_adapt.object_model.interface.measures import HazardType, ImpactType
 from flood_adapt.object_model.interface.strategies import IStrategy, StrategyModel
 from flood_adapt.object_model.measure_factory import (
