+++ tests/test_object_model/test_scenarios.py
@@ -2,7 +2,6 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from flood_adapt.dbs_classes.path_builder import TopLevelDir, db_path
 from flood_adapt.dbs_controller import Database
 from flood_adapt.object_model.direct_impact.impact_strategy import ImpactStrategy
 from flood_adapt.object_model.direct_impact.socio_economic_change import (
@@ -12,6 +11,10 @@ from flood_adapt.object_model.direct_impacts import DirectImpacts
 from flood_adapt.object_model.hazard.hazard import Hazard
 from flood_adapt.object_model.hazard.hazard_strategy import HazardStrategy
 from flood_adapt.object_model.hazard.physical_projection import PhysicalProjection
+from flood_adapt.object_model.interface.database import (
+    TopLevelDir,
+    db_path,
+)
 from flood_adapt.object_model.interface.events import RainfallModel, TideModel
 from flood_adapt.object_model.interface.site import SCSModel, Site
 from flood_adapt.object_model.io.unitfulvalue import UnitfulLength
